Refactor DialogTests into partial classes by concern
Split DialogTests.cs into Layout, Drawing, and Generic partial class files for better organization. Updated event handler lambdas to use discards for unused parameters per code style. No changes to test logic; all tests preserved and coverage maintained.
